+++ src/File/Count.php
@@ -80,13 +80,16 @@ class Count extends AbstractValidator
      */
     public function __construct($options = null)
     {
-        if (is_string($options) || is_numeric($options)) {
-            $options = ['max' => $options];
+        if (1 < func_num_args()) {
+            $args = func_get_args();
+            $options = [
+                'min' => array_shift($args),
+                'max' => array_shift($args),
+            ];
         }
 
-        if (1 < func_num_args()) {
-            $options['min'] = func_get_arg(0);
-            $options['max'] = func_get_arg(1);
+        if (is_string($options) || is_numeric($options)) {
+            $options = ['max' => $options];
         }
 
         parent::__construct($options);
@@ -111,11 +114,11 @@ class Count extends AbstractValidator
      */
     public function setMin($min)
     {
-        if (is_array($min) and isset($min['min'])) {
+        if (is_array($min) && isset($min['min'])) {
             $min = $min['min'];
         }
 
-        if (!is_string($min) and !is_numeric($min)) {
+        if (! is_numeric($min)) {
             throw new Exception\InvalidArgumentException('Invalid options to validator provided');
         }
 
@@ -149,11 +152,11 @@ class Count extends AbstractValidator
      */
     public function setMax($max)
     {
-        if (is_array($max) and isset($max['max'])) {
+        if (is_array($max) && isset($max['max'])) {
             $max = $max['max'];
         }
 
-        if (!is_string($max) and !is_numeric($max)) {
+        if (! is_numeric($max)) {
             throw new Exception\InvalidArgumentException('Invalid options to validator provided');
         }