+++ test/File/HashTest.php
@@ -10,6 +10,7 @@
 namespace ZendTest\Validator\File;
 
 use Zend\Validator\File;
+use ReflectionProperty;
 
 /**
  * Hash testbed
@@ -146,7 +147,10 @@ class HashTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['12345' => 'crc32', '12344' => 'crc32'], $validator->getHash());
 
         $validator->addHash(['12321', '12121']);
-        $this->assertEquals(['12345' => 'crc32', '12344' => 'crc32', '12321' => 'crc32', '12121' => 'crc32'], $validator->getHash());
+        $this->assertEquals(
+            ['12345' => 'crc32', '12344' => 'crc32', '12321' => 'crc32', '12121' => 'crc32'],
+            $validator->getHash()
+        );
     }
 
     /**
@@ -177,5 +181,63 @@ class HashTest extends \PHPUnit_Framework_TestCase
 
         $this->assertFalse($validator->isValid($filesArray));
         $this->assertArrayHasKey(File\Hash::NOT_FOUND, $validator->getMessages());
+    }
+
+    public function invalidHashTypes()
+    {
+        return [
+            'null'       => [null],
+            'true'       => [true],
+            'false'      => [false],
+            'zero'       => [0],
+            'int'        => [1],
+            'zero-float' => [0.0],
+            'float'      => [1.1],
+            'object'     => [(object) []],
+        ];
+    }
+
+    /**
+     * @dataProvider invalidHashTypes
+     */
+    public function testAddHashRaisesExceptionForInvalidType($value)
+    {
+        $validator = new File\Hash('12345');
+        $this->setExpectedException('Zend\Validator\Exception\InvalidArgumentException', 'False parameter given');
+        $validator->addHash($value);
+    }
+
+    public function testAddHashRaisesExceptionWithInvalidAlgorithm()
+    {
+        $validator = new File\Hash('12345');
+        $algorithm = 'foobar123';
+        $options   = ['algorithm' => $algorithm];
+        $this->setExpectedException(
+            'Zend\Validator\Exception\InvalidArgumentException',
+            sprintf("Unknown algorithm '%s'", $algorithm)
+        );
+        $validator->addHash($options);
+    }
+
+    public function testIsValidRaisesExceptionForArrayValueNotInFilesFormat()
+    {
+        $validator = new File\Hash('12345');
+        $value     = ['foo' => 'bar'];
+        $this->setExpectedException(
+            'Zend\Validator\Exception\InvalidArgumentException',
+            'Value array must be in $_FILES format'
+        );
+        $validator->isValid($value);
+    }
+
+    public function testConstructorCanAcceptAllOptionsAsDiscreteArguments()
+    {
+        $algorithm = 'md5';
+        $validator = new File\Hash('12345', $algorithm);
+
+        $r = new ReflectionProperty($validator, 'options');
+        $r->setAccessible(true);
+        $options = $r->getValue($validator);
+        $this->assertSame($algorithm, $options['algorithm']);
     }
 }